Integrate 106x-specific crate into iomuxc crate
Rather than having a separate crate per chip implementation, I'm
proposing that we have a single, imxrt-iomuxc crate. We expose
chip-specific pads behind feature flags. I know this is the opposite
of what I originally proposed, but hear me out.

We won't require a feature flag, but a feature flags will be needed
to access processor-specific pads. The interfaces are the only thing
that we expose if we don't use a feature flag, so that's the same as
using the imxrt-iomuxc crate by itself. Adding more and more feature
flags is the same as depending on more and more crates. If we expand
that through a user's dependency graph, it ends up being equivalent.

But, managing a single crate with multiple feature flags is easier
than managing multiple crates that effectively represent features.
We will still compile all of the tests without feature flags. And,
we can document everything by specifying '--all-features'. Finally,
we can expose a simpler interface without the "if it's not
documented, it's not public" clauses.