iomuxc: eagerly bump version to 0.2
The feature-flag renaming is a breaking change. This commit bumps the
imxrt-iomuxc commit to 0.2. The commit updates the HAL to use the new
version.

The HAL now re-exports a 0.2 imxrt-iomuxc API. Although the symbol
names exported from the HAL have not changed, the compiler will still
see the types as being incompatible. This means that the HAL has a
breaking change. I'm relying on other incoming work in the project
(#89) to bump the version for me.

Or, we hold off on depending on the imxrt-iomuxc 0.2 release, and
nothing breaks.