Favor macros for pin implementations
The SAI pin implementations introduced macros to define the 1060 SAI
pins. This commit doubles down on the approach by making macros the
default way to define pins. This makes it easier to study pin
implementations at a glance, and keeps consistency throughout the code.

There's extra spacing in between the macros. The goal is to make a diff
easier to study. I made these changes simply with a multi cursor, and
spot checked parts of the changes. All looked OK. The next commit
collapses the spacing.

Cherry-picked from main branch, with SCREAMING symbol renames.