Add uSDCH interface, 1060 pins
Checked 1060 pins against table 10-1 in the reference manual. The pins
are sufficient to express the uSDHC peripheral on a Teensy 4.1. It does
not include all possible pins for all uSDHC peripherals, nor does it
include all signals (like RESET_B). These can come later. Verified that
these pins do not include DAISY registers. Empty daisy module is a
placeholder for the future.

This Pin trait is the first to introduce pin configurations for the
signals. This is great, provided that it makes sense for all boards and
driver implementations. We could consider expanding this concept to
other Pin traits in the future.

This is merge commit edbb3d2, forward ported onto the main branch.