Refactor GPIO pin trait
Some pins on the 1170 support two different GPIO modules. We couldn't
express that with the previous trait. This commit moves the module
number into a const generic argument so that the same pad can implement
multiple pin traits.

This removes the gpio::Pin trait bound on ADC, for convenience. It
doesn't seem like we need this bound; see inline notes.