Revise README, API documentation
Revise the README following some of the ideas in #2. This commit drops
the design documentation that influenced today's crates. We can still
view it through the project history.

Consolidate the API documentation in the package documentation, dropping
the duplicated example from the 1010 module. Update the documentation
examples to remove some unnecessary `unsafe` blocks.

Closes #2.