Hard-code alternate for adc::prepare
Using '5' for the GPIO alternate. Troubleshooting revealed that this was
necessary for a working ADC input, disputing the inline comment. It's
simpler to hard-code the alt than to put the generic constraints on
users.

Even though the `configure` call isn't working for the 1170, I believe
'5' is also the correct alternate to set for the 1170 LPADC. See the new
inline comment for the consideration.