Fix superficial differences between 1050 and 1060
These three modules are nearly the same between the 1050 and 1060.
There's only some superficial naming differences. Since the 1060 is
established, let's prefer those names (even if the 1050 names are better
in some cases).

I produced this diff by copying the three modules from the 1060
directory. This is an intermediate diff to show that these peripherals
are effectively the same.